Moved tree related structures to treeviewer package
